feat(backend): add ticket logic (#382)
Gno+backend ticket creation / validation

### Flow

- A ticket secret is 32 bytes of secure random, it is used to seed an
ed25519 private key (for example using go's
`crypto/ed25517.NewKeyFromSeed`) and is encoded to string in
base64url-nopadding. It is created by the backend on event
participation.
- This ticket secret is stored in db and sent to the user by email in
the form of a QR code.
- The ticket QR should also be accessible on the web ui so we need a
backend call to retrieve an user tickets secrets when they are
logged-in.
- The ticket pubkey is the corresponding ed215519 pubkey encoded to
string in base64url-nopadding
- When a ticket is created, the ticket pubkey is posted on-chain.
- To validate a ticket, the gatekeeper goes to the gatekeeper page of
the event and scans the ticket secret and derive it's pubkey to check
chain status of the ticket. If this ticket is already used, entrance is
denied. If it's unused, it creates a signature of it's gno address using
the private key and attempts to send a `MsgCheckin` containing the
ticket pubkey and this signature. If the transaction is settled, the
gatekeeper can let the participant in.

Signing the address of the current gatekeeper prevent a compromised
gatekeeper from trying to front-run check-in txs and dos the event
entrance.

### Changes

##### Gno

- In `Event` type, add a new avl tree containing tickets with the ticket
pubkey as key
- Edit the `gno.land/p/zenao/events.AddParticipant` action to also take
a ticket pubkey and add this ticket to the tickets tree on execution. It
panics if the ticket has already been added.
- Create the `gno.land/p/zenao/events.Checkin` action, usable by
gatekeepers only. It takes a ticket pubkey and signature as input. It
marks this ticket as used. It panics if the signature is invalid or the
ticket does not exists or it is already used.
- Refacto event init to easily add new resources by roles
- Add `backend genticket` cli subcommand to generate a testing ticket
- Add ticket creation and checkin test

##### Backend

- Add `Ticket` type in `zeni` package that is a wrapper around the
secret key and exposes methods to get the ticket pubkey, secret and
signature easily as well as constructors to create a new secure ticket
or restore a ticket from it's secret.
- Adapt `Participate` call in service, chain and db to create a new
ticket and store it
- Add `GetEventBuyerTickets` db method to retrieve the tickets bought
for an event by a particular user
- Adapt `sync-chain` command to get the ticket from the db during
participation replay
- Change `SoldTicket.UserID string` field to `SoldTicket.BuyerID uint`
to reflect that an user can buy multiple tickets and is not necessarily
who's going to use it.
- Add `GetEventTickets` zenao api endpoint to retrieve an user tickets
for a particular event from web app

---------

Signed-off-by: Norman <norman@samourai.coop>