feat(back): add a way to create with many organizers & to edit it (#484)
This PR add fields ``organizers`` in Create & Edit Event endpoints:
- A user cannot remove himself as an organizer, so don't put the caller
as an organizer because it will be added anyway and if you add it in the
array it will raise an error as a duplicate organizer.
- I fixed front-end to work but it only display the first organizer for
now & add a little translation with i18n where it was missing
- Be careful, once you put someone organizer he can kick you there is no
difference between a creator & an organizer especially on-chain where
the "creator" entity does not exists
- the event registry now index for all organizers & update the index
when the organizers change
- gnoweb display should be working
- I did not add new gno unit test to ensure everything works, i just
adapted the old one to pass
- I tested the PR by creating an event with 2 organizer and edit with 3
organizer including one that was already organizer.

P.S: The creation/edition can be very long depending on the number of
users, especially if we have to create them a user realm & clerk account

---------

Signed-off-by: Norman <norman@samourai.coop>
Co-authored-by: Norman <norman@samourai.coop>