fix(gentxs): set the created time when save update the records (#706)
The problem is we don't keep historic of join and leave but just the
latest status.
When generating txs if some action were made before you leave and you
re-join after the organization it will looks like in database you did
the action when you was not part of the organization.

It's a problem and a simple solution for now is to keep the created_at
field at first time you get the role in the organization. if you leave
and re-join, the gentxs will make it like you never left.

The best solution would be to be able to replay all txs.

set created_at to not change after creation
see:
https://gorm.io/docs/models.html?utm_source=chatgpt.com#Field-Level-Permission

Also the save method overwrite the created_at value, it's a way to avoid
this:

for example it was causing the issue of gentxs in staging by having 0
value created_at for users that left and re-join a community:

<img width="860" height="184" alt="image"
src="https://github.com/user-attachments/assets/d66dd891-ed6a-42bc-b804-3af17b2d374c"
/>

---------

Co-authored-by: n0izn0iz <n0izn0iz@users.noreply.github.com>