fix(front): Event page hydration mismatch (#878)
Bugfix proposal combining #871 & #743 to erase hydration mismatch and
stabilize CI.

**Issue**

Hydration mismatch errors are cause by race-conditions between slow
client hydration and async effects. It can look like a "timing issue"
but it's not.

Because if the network or React rendering is slightly slower or faster,
the mismatch appears or disappears.

But the *real* cause isn’t the speed — it’s that **the server and client
are rendering different initial states**.

**Changes**

- Frontend event page refactoring: Removing SSG, Lazy-Loading on
critical components

From #743 

- improve event details load time
- add loading.tsx for event details routes
- rearchitecture to support showing an event skeleton during loading
- add various skeletons and suspenses

---------

Signed-off-by: Norman <norman@samourai.coop>
Co-authored-by: Norman <norman@samourai.coop>
Co-authored-by: n0izn0iz <n0izn0iz@users.noreply.github.com>
Co-authored-by: MikaelVallenet <mikael.vallenetpro@gmail.com>