chore: rollback staging database (#877)
I added migration file i'd like to rollback while working on SoT.
There is 2 ways of doing this:
- rollback the migrations but since i dropped data and modify the
foreign key relationship it's a bit though to do this and we will still
miss a lot of data.
- Use a backup, before migrating i did a backup, in turso we cannot
rename database so i had to use a new database that's why this PR change
the name to use the backup instead of the original one.