fix(front): improve fallback communities list page (#903)
**Error occuring**
```
Error: An unknown Component is an async Client Component. Only Server Components can be async at the moment. This error is often caused by accidentally adding `'use client'` to a module that was originally written for the server.
```

Full trace here:
https://sentry.samourai.pro/organizations/sentry/issues/1004/?alert_rule_id=3&alert_type=issue&environment=production&notification_uuid=48e95f32-9979-449d-85ff-3a832dfccec2&project=3

Not sure of the exact source of the problem: while investigating, I
could not find any async Client Component used in the route
`/communities`.

As problem occurs on on after click on loading elements, I suggest for
the moment to expand Suspense boudaries to get a more precise output.

**Changes**

- add `loading.tsx` on `/communities`
- add more `Suspense` boundaries & fallbacks 
- update i18n

---------

Co-authored-by: MikaelVallenet <mikael.vallenetpro@gmail.com>